+++ rstar/src/primitives/mod.rs
@@ -1,12 +1,12 @@
 //! Contains primitives ready for insertion into an r-tree.
 
-mod geom_with_cached_envelope;
+mod cached_envelope;
 mod geom_with_data;
 mod line;
 mod point_with_data;
 mod rectangle;
 
-pub use self::geom_with_cached_envelope::GeomWithCachedEnvelope;
+pub use self::cached_envelope::CachedEnvelope;
 pub use self::geom_with_data::GeomWithData;
 pub use self::line::Line;
 pub use self::point_with_data::PointWithData;