+++ rstar/CHANGELOG.md
@@ -1,7 +1,7 @@
 # UNRELEASED
 
 ## Added
-- Add `GeomWithCachedEnvelope` combinator which simplifies memoizing envelope computations. ([PR](https://github.com/georust/rstar/pull/118))
+- Add `CachedEnvelope` combinator which simplifies memoizing envelope computations. ([PR](https://github.com/georust/rstar/pull/118))
 
 # 0.10.0